Move AbstractReporterBenchmark to jmh module
