+++ core/build.gradle
@@ -19,6 +19,12 @@
 // THE SOFTWARE.
 
 description = 'Interfaces and utilities to report metrics to M3'
+apply from: 'jmhFixtures.gradle'
+
+//dependencies {
+//    jmhCompile('org.openjdk.jmh:jmh-core:1.27')
+//    jmhCompile('org.openjdk.jmh:jmh-generator-annprocess:1.27')
+//}
 
 sourceSets {
     jmh {
@@ -29,12 +35,7 @@ sourceSets {
     }
 }
 
-dependencies {
-    compile('org.openjdk.jmh:jmh-core:1.27')
-    compile('org.openjdk.jmh:jmh-generator-annprocess:1.27')
-}
-
-task jmh(type: JavaExec, dependsOn: jmhClasses) {
+task runJmhTests(type: JavaExec, dependsOn: jmhClasses) {
     main = 'org.openjdk.jmh.Main'
     classpath = sourceSets.jmh.compileClasspath + sourceSets.jmh.runtimeClasspath
 }
