+++ statsd/build.gradle
@@ -21,10 +21,8 @@
 description = 'tally StatsD reporter'
 
 dependencies {
-    compile project(':tally-core')
     compile('com.datadoghq:java-dogstatsd-client:2.3')
-    compile('org.openjdk.jmh:jmh-core:1.27')
-    compile('org.openjdk.jmh:jmh-generator-annprocess:1.27')
+    compile project(path: ':tally-core', configuration: 'jmhFixturesUsageCompile')
 }
 
 sourceSets {
@@ -36,7 +34,7 @@ sourceSets {
     }
 }
 
-task jmh(type: JavaExec, dependsOn: jmhClasses) {
+task runJmhTests(type: JavaExec, dependsOn: jmhClasses) {
     main = 'org.openjdk.jmh.Main'
     classpath = sourceSets.jmh.compileClasspath + sourceSets.jmh.runtimeClasspath
 }
